pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-assist-ops-env"
version = "0.1.0"
description = "Assist Ops environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.1",
    "fastapi",
    "uvicorn",
    "openai>=1.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "assist_ops_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["assist_ops_env", "assist_ops_env.server"]
package-dir = { "assist_ops_env" = ".", "assist_ops_env.server" = "server" }
